requirements.txt
# Python 3.11+

# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Telegram Bot
aiogram>=3.17.0,<4.0.0

# Web Panel
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
waitress>=3.0.0
werkzeug>=3.0.0

# Async HTTP and Database
aiohttp>=3.9.0
aiosqlite>=0.19.0

# Machine Learning
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# NLP and Transformers
transformers[torch]>=4.35.0

# Text Processing
emoji>=2.8.0
regex>=2023.10.3

# OpenAI Integration (optional)
openai>=1.0.0
